used vec<chars> to represent java string

+++ src/token.rs
@@ -52,11 +52,11 @@ pub struct Token {
     token_type: TokenType,
     lexeme: String,
     literal: Option<Literal>,
-    line: u32,
+    line: usize,
 }
 
 impl Token {
-    pub fn new(token_type: TokenType, lexeme: String, literal: Option<Literal>, line: u32) -> Self {
+    pub fn new(token_type: TokenType, lexeme: String, literal: Option<Literal>, line: usize) -> Self {
         Self {
             token_type,
             lexeme,
